feat: move python files used in snakemake to the scripts folder
