+++ setup.cfg
@@ -1,3 +1,4 @@
 [coverage:run]
 omit =
     */test/*
+    setup.py
